Create reminders.json
Structure of the content:

This code will accept the contents of an array [ ].  Each object is enclosed in curly braces { } and represents a single reminder typed in by the user.  Each reminder that has been typed in by the user contains three properties, "text", "time", "notes".  Each property is a unique key-value pair, separated by a colon.  

The key is a string that represents the name of the property.  The value is the data associated with that property. i.e. the value accepted must be in text format.  The objects in the array are separated by a comma.